Get all tests to run for a first time...
