+++ src/sqooler/spoolers.py
@@ -2,6 +2,23 @@
 This module contains the code for the Spooler classes and its helpers.
 """
 
+import os
+from collections.abc import Callable
+from typing import Optional, Type, Any
+from time import sleep
+
+from pydantic import ValidationError, BaseModel, Field
+from icecream import ic
+
+from .schemes import (
+    BackendConfigSchemaIn,
+    ExperimentDict,
+    ResultDict,
+    StatusMsgDict,
+    GateDict,
+    LabscriptParams,
+)
+
 
 class Spooler:
     """
@@ -313,6 +330,8 @@ class LabscriptSpooler(Spooler):
         device_config: Type[BaseModel],
         n_wires: int,
         remote_client: Any,  # it would be really nice to fix this type
+        labscript_params: LabscriptParams,
+        run: Any,  # it would be really nice to fix this type
         description: str = "",
         n_max_shots: int = 1000,
         version: str = "0.0.1",
@@ -323,7 +342,13 @@ class LabscriptSpooler(Spooler):
         operational: bool = True,
     ):
         """
-        The constructor of the class.
+        The constructor of the class. The  arguments are the same as for the Spooler
+        class with two additions.
+
+        Args:
+            remote_client: The remote client that is used to connect to the labscript server.
+            labscript_params: The parameters that are used to generate the folder for the shots.
+            run: The run object that is used to execute the labscript file.
         """
         super().__init__(
             ins_schema_dict,
@@ -339,6 +364,8 @@ class LabscriptSpooler(Spooler):
             operational,
         )
         self.remote_client = remote_client
+        self.labscript_params = labscript_params
+        self.run = run
 
     def add_job(
         self, json_dict: dict, status_msg_dict: StatusMsgDict
@@ -377,8 +404,18 @@ class LabscriptSpooler(Spooler):
                     self._modify_shot_output_folder(job_id + "/" + str(exp))
 
                     # Here we generate the ciruit
-                    result_draft["results"].append(self.gen_circuit(exp_dict, job_id))
-
+                    try:
+                        result_draft["results"].append(
+                            self.gen_circuit(exp_dict, job_id)
+                        )
+                    except FileNotFoundError as err:
+                        ic(err)
+                        error_message = str(err)
+                        status_msg_dict.detail += "; Failed to generate labscript file."
+                        status_msg_dict.error_message += f"; Failed to generate labscript file. Error: {error_message}"
+                        status_msg_dict.status = "ERROR"
+                        result_dict = ResultDict(**result_draft)
+                        return result_dict, status_msg_dict
                 status_msg_dict.detail += "; Passed json sanity check; Compilation done. \
                     Shots sent to solver."
                 status_msg_dict.status = "DONE"
@@ -411,18 +448,155 @@ class LabscriptSpooler(Spooler):
         result_dict = ResultDict(**result_draft)
         return result_dict, status_msg_dict
 
-    def _modify_shot_output_folder(self, new_dir: str) -> None:
+    def _modify_shot_output_folder(self, new_dir: str) -> str:
         """
         I am not sure what this function does.
 
         Args:
             new_dir: The new directory under which we save the shots.
+
+        Returns:
+            The path to the new directory.
         """
+
+        # we should simplify this at some point
         defaut_shot_folder = str(self.remote_client.get_shot_output_folder())
-        print(f"Default shot folder: {defaut_shot_folder}")
+
         modified_shot_folder = (defaut_shot_folder.rsplit("\\", 1)[0]) + "/" + new_dir
-        print(f"Modified shot folder: {modified_shot_folder}")
+        # suggested better emthod whcih works the same way on all platforms
+        # modified_shot_folder = os.path.join(
+        #    os.path.dirname(defaut_shot_folder), new_dir
+        # )
         self.remote_client.set_shot_output_folder(modified_shot_folder)
+        return modified_shot_folder
+
+    def gen_circuit(self, json_dict: dict, job_id: str) -> ExperimentDict:
+        """
+        This is the main script that generates the labscript file.
+
+        Args:
+            json_dict: The dictionary that contains the instructions for the circuit.
+            job_id: The user id of the user that is running the experiment.
+
+        Returns:
+            The path to the labscript file.
+        """
+
+        # parameters for the function
+        EXP_SCRIPT_FOLDER = self.labscript_params.exp_script_folder
+
+        # local files
+        HEADER_PATH = f"{EXP_SCRIPT_FOLDER}/header.py"
+        REMOTE_EXPERIMENTS_PATH = f"{EXP_SCRIPT_FOLDER}/remote_experiments"
+
+        exp_name = next(iter(json_dict))
+        ins_list = json_dict[next(iter(json_dict))]["instructions"]
+        n_shots = json_dict[next(iter(json_dict))]["shots"]
+
+        globals_dict = {
+            "job_id": "guest",
+            "shots": 4,
+        }
+        globals_dict["shots"] = list(range(n_shots))
+        globals_dict["job_id"] = job_id
+
+        self.remote_client.set_globals(globals_dict)
+        script_name = f"experiment_{globals_dict['job_id']}.py"
+        exp_script = os.path.join(REMOTE_EXPERIMENTS_PATH, script_name)
+        ins_list = json_dict[next(iter(json_dict))]["instructions"]
+        print(f"File path: {exp_script}")
+        code = ""
+        # this is the top part of the script it allows us to import the
+        # typical functions that we require for each single sequence
+        # first have a look if the file exists
+        if not os.path.exists(HEADER_PATH):
+            ic(HEADER_PATH)
+            raise FileNotFoundError("Header file not found.")
+
+        with open(HEADER_PATH, "r", encoding="UTF-8") as header_file:
+            code = header_file.read()
+
+        # add a line break to the code
+        code += "\n"
+
+        try:
+            with open(exp_script, "w", encoding="UTF-8") as script_file:
+                script_file.write(code)
+        except:
+            print("Something wrong. Does file path exists?")
+
+        for inst in ins_list:
+            # we can directly use the function name as we have already verified
+            # that the function exists in the `add_job` function
+            func_name = inst[0]
+            params = "(" + str(inst[1:])[1:-1] + ")"
+            code = "Experiment." + func_name + params + "\n"
+
+            # we should add proper error handling here
+            # pylint: disable=bare-except
+            try:
+                with open(exp_script, "a", encoding="UTF-8") as script_file:
+                    script_file.write(code)
+            except:
+                print("Something wrong. Does file path exists?")
+
+        code = "Experiment.final_action()" + "\n" + "stop(Experiment.t+0.1)"
+        # pylint: disable=bare-except
+        try:
+            with open(exp_script, "a", encoding="UTF-8") as script_file:
+                script_file.write(code)
+        except:
+            print("Something wrong. Does file path exists?")
+        self.remote_client.set_labscript_file(
+            exp_script
+        )  # CAUTION !! This command only selects the file. It does not generate it!
+
+        # be careful. This is not a blocking command
+        self.remote_client.engage()
+
+        # now that we have engaged the calculation we need to wait for the
+        # calculation to be done
+
+        # we need to get the current shot output folder
+        current_shot_folder = self.remote_client.get_shot_output_folder()
+        ic(current_shot_folder)
+        # we need to get the list of files in the folder
+        ic("Get hdf5 files")
+        hdf5_files = get_file_queue(current_shot_folder)
+
+        ic("Wait for them")
+        ic(hdf5_files)
+        ic(current_shot_folder)
+        # we need to wait until we have the right number of files
+        while len(hdf5_files) < n_shots:
+            ic("Wait a bit")
+            sleep(self.labscript_params.t_wait)
+            hdf5_files = get_file_queue(current_shot_folder)
+
+        shots_array = []
+        # once the files are there we can read them
+        for file in hdf5_files:
+            this_run = self.run(current_shot_folder + "/" + file)
+            got_nat = False
+            n_tries = 0
+            # sometimes the file is not ready yet. We need to wait a bit
+            while not got_nat and n_tries < 15:
+                # the exception is raised if the file is not ready yet
+                # it is broadly defined within labscript so we cannot do anything about
+                # it here.
+                # pylint: disable=W0718
+                try:
+                    print(this_run.get_results("/measure", "nat"))
+                    # append the result to the array
+                    shots_array.append(this_run.get_results("/measure", "nat"))
+                    got_nat = True
+                except Exception as exc:
+                    print(exc)
+                    sleep(self.labscript_params.t_wait)
+                    n_tries += 1
+
+        exp_sub_dict = create_memory_data(shots_array, exp_name, n_shots)
+        return exp_sub_dict
 
 
 def gate_dict_from_list(inst_list: list) -> GateDict:
@@ -438,3 +612,52 @@ def gate_dict_from_list(inst_list: list) -> GateDict:
     """
     gate_draft = {"name": inst_list[0], "wires": inst_list[1], "params": inst_list[2]}
     return GateDict(**gate_draft)
+
+
+def get_file_queue(dir_path: str) -> list[str]:
+    """
+    A function that returns the list of files in the directory.
+
+    Args:
+        dir_path: The path to the directory.
+
+    Returns:
+        A list of files in the directory. It excludes directories.
+
+    Raises:
+        ValueError: If the path is not a directory.
+    """
+
+    # make sure that the path is an existing directory
+    if not os.path.isdir(dir_path):
+        raise ValueError(f"The path {dir_path} is not a directory.")
+    files = [
+        file
+        for file in os.listdir(dir_path)
+        if os.path.isfile(os.path.join(dir_path, file))
+    ]
+    return files
+
+
+def create_memory_data(
+    shots_array: list, exp_name: str, n_shots: int
+) -> ExperimentDict:
+    """
+    The function to create memory key in results dictionary
+    with proprer formatting.
+    """
+    exp_sub_dict: dict = {
+        "header": {"name": "experiment_0", "extra metadata": "text"},
+        "shots": 3,
+        "success": True,
+        "data": {"memory": None},
+    }
+
+    exp_sub_dict["header"]["name"] = exp_name
+    exp_sub_dict["shots"] = n_shots
+    memory_list = [
+        str(shot).replace("[", "").replace("]", "").replace(",", "")
+        for shot in shots_array
+    ]
+    exp_sub_dict["data"]["memory"] = memory_list
+    return ExperimentDict(**exp_sub_dict)
